qty per dose and frequency list updated

+++ app/src/main/java/com/heartcare/agni/data/server/model/prescription/prescriptionresponse/Medication.kt
@@ -11,8 +11,8 @@ data class Medication(
     val frequency: Int,
     val medFhirId: String,
     val note: String?,
-    val qtyPerDose: Int,
-    val qtyPrescribed: Int,
+    val qtyPerDose: Double,
+    val qtyPrescribed: Double,
     val timing: String?,
     val brandName: String?,
     val doseFormCode: String?
